feat: add liquibase and скрипты на создание и заполнение
разбил конфигурационный файл.
