+++ src/main/java/ru/clevertec/ecl/config/JacksonConfig.java
@@ -1,0 +1,16 @@
+package ru.clevertec.ecl.config;
+
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.springframework.context.annotation.Bean;
+import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+import org.springframework.stereotype.Component;
+
+@Component
+public class JacksonConfig {
+
+    @Bean
+    public Jackson2ObjectMapperBuilder jackson2ObjectMapperBuilder() {
+        return new Jackson2ObjectMapperBuilder()
+                .modules(new JavaTimeModule());
+    }
+}
