+++ src/main/java/ru/clevertec/ecl/config/LiquibaseConfig.java
@@ -1,0 +1,24 @@
+package ru.clevertec.ecl.config;
+
+import liquibase.integration.spring.SpringLiquibase;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.stereotype.Component;
+
+import javax.sql.DataSource;
+
+@Component
+public class LiquibaseConfig {
+
+    @Value("${change-log}")
+    private String changeLog;
+
+    @Bean
+    public SpringLiquibase liquibase(DataSource dataSource) {
+        SpringLiquibase liquibase = new SpringLiquibase();
+        liquibase.setDataSource(dataSource);
+        liquibase.setChangeLog(changeLog);
+
+        return liquibase;
+    }
+}
