+++ src/main/java/ru/clevertec/ecl/config/SpringConfig.java
@@ -1,7 +1,5 @@
 package ru.clevertec.ecl.config;
 
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
 import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
 import org.springframework.context.annotation.Bean;
@@ -10,54 +8,17 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
 import org.springframework.core.io.ClassPathResource;
-import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.orm.hibernate5.HibernateTransactionManager;
-import org.springframework.orm.hibernate5.LocalSessionFactoryBean;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import javax.sql.DataSource;
 import java.util.Objects;
 import java.util.Properties;
 
 @EnableWebMvc
 @Configuration
-@EnableTransactionManagement
 @ComponentScan("ru.clevertec")
 @PropertySource("classpath:application.yml")
 public class SpringConfig implements WebMvcConfigurer {
-
-//    private Environment environment;
-
-//    @Value("${enabled}")
-//    private boolean enabled;
-//
-//    @Value("${change-log}")
-//    private String changeLog;
-
-    @Value("${url}")
-    private String url;
-
-    @Value("${user}")
-    private String userName;
-
-    @Value("${password}")
-    private String password;
-
-    @Value("${driver}")
-    private String driver;
-
-    @Value("${hibernate.dialect}")
-    private String dialect;
-
-    @Value("${hibernate.show_sql}")
-    private String showSql;
-
-    @Value("${hibernate.format_sql}")
-    private String formatSql;
 
     @Bean
     public static BeanFactoryPostProcessor beanFactoryPostProcessor() {
@@ -66,66 +27,7 @@ public class SpringConfig implements WebMvcConfigurer {
         yaml.setResources(new ClassPathResource("application.yml"));
         Properties yamlObject = Objects.requireNonNull(yaml.getObject(), "Yaml not found.");
         configure.setProperties(yamlObject);
+
         return configure;
     }
-
-    @Bean
-    public Jackson2ObjectMapperBuilder jackson2ObjectMapperBuilder() {
-        return new Jackson2ObjectMapperBuilder()
-                .modules(new JavaTimeModule());
-    }
-
-    @Bean
-    public DataSource dataSource() {
-        DriverManagerDataSource dataSource = new DriverManagerDataSource();
-
-        dataSource.setDriverClassName(driver);
-        dataSource.setUrl(url);
-        dataSource.setUsername(userName);
-        dataSource.setPassword(password);
-
-        return dataSource;
-    }
-
-    @Bean
-    public LocalSessionFactoryBean sessionFactory() {
-        LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
-        sessionFactory.setDataSource(dataSource());
-        sessionFactory.setPackagesToScan("ru.clevertec.ecl.entity");
-        sessionFactory.setHibernateProperties(hibernateProperties());
-
-        return sessionFactory;
-    }
-
-    @Bean
-    public PlatformTransactionManager hibernateTransactionManager() {
-        HibernateTransactionManager transactionManager = new HibernateTransactionManager();
-        transactionManager.setSessionFactory(sessionFactory().getObject());
-
-        return transactionManager;
-    }
-
-    private Properties hibernateProperties() {
-        Properties properties = new Properties();
-        properties.put("hibernate.dialect", dialect);
-        properties.put("hibernate.show_sql", showSql);
-        properties.put("hibernate.format_sql", formatSql);
-
-        return properties;
-    }
-//    @PostConstruct
-//    private void configLiquibase() {
-//        try {
-//
-//            if (enabled) {
-//                Database database = DatabaseFactory.getInstance()
-//                        .findCorrectDatabaseImplementation(new JdbcConnection(getConnection()));
-//                Liquibase liquibase = new Liquibase(changeLog, new ClassLoaderResourceAccessor(), database);
-//                liquibase.update(new Contexts(), new LabelExpression());
-//            }
-//        } catch (Exception e) {
-//            e.printStackTrace();
-//        }
-//    }
-
 }
