+++ src/main/resources/db/changelog/house.sql
@@ -1,0 +1,27 @@
+INSERT INTO public.house (id, uuid, area, country, city, street, number, create_date)
+VALUES (1, '99efee95-2f1c-459e-b97c-509f7399aa01', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-12 23:29:04.595808'),
+       (3, 'd7d5c085-d470-4f4b-8d3c-f04a1305d4ea', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-13 01:26:12.411684'),
+       (4, '99efee96-2f1c-459e-b97c-509f7399aa02', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-12 23:29:04.595000'),
+       (5, '99efee96-2f1c-459e-b97c-509f7399aa03', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-12 23:29:04.595000'),
+       (8, '633677d4-31a4-4728-b298-bfd8bbae8445', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-14 20:10:17.985000'),
+       (9, '633677d4-31a4-4728-b298-bfd8bbae8446', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-14 20:10:17.985000'),
+       (10, '633677d4-31a4-4728-b298-bfd8bbae8447', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-14 20:10:17.985000'),
+       (14, '99efee95-2f1c-459e-b97c-509f7399aa02', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-12 23:29:04.595000'),
+       (16, '99efee95-2f1c-459e-b97c-509f7399aa03', 'Homelska area', 'Belarus country', 'Homel city', 'Sovet street',
+        123, '2024-01-12 23:29:04.595000'),
+       (17, '633677d4-31a4-4728-b298-bfd8bbae8448', 'Some area', '1 country', 'Some city', 'Some street', 123,
+        '2024-01-14 20:10:17.985000'),
+       (21, '633677d4-32a4-4728-b298-bfd8bbae8445', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-14 20:10:17.985000'),
+       (22, '99efee95-2f1c-459e-b97c-519f7399aa03', 'Homelska area', 'Belarus country', 'Homel city', 'Sovet street',
+        123, '2024-01-12 23:29:04.595000'),
+       (23, '99efee96-2f1c-459e-b97c-609f7399aa03', 'Some area', 'Some country', 'Some city', 'Some street', 123,
+        '2024-01-12 23:29:04.595000');
