+++ src/main/resources/db/changelog/person.sql
@@ -1,0 +1,12 @@
+INSERT INTO public.person (id, uuid, name, surname, sex, passport_series, passport_number, create_date, update_date,
+                           house_id)
+VALUES (1, 'd8b6eda1-2ac7-4190-8523-389b3cccffa9', 'John', 'Doe', 'Male', 'HB', 3, '2024-01-12 23:29:04.595808',
+        '2024-01-12 23:29:04.595808', 1),
+       (3, '3cd20b71-6381-4658-9711-f834f1c3373a', 'John', 'Doe', 'Male', 'HB', 1, '2024-01-13 01:26:12.411684',
+        '2024-01-13 01:26:12.411684', 3),
+       (9, 'd8b6eda1-2ac7-4190-8523-389b3cccffa8', 'John', 'Doe', 'Male', 'HB', 4, '2024-01-12 23:29:04.595000',
+        '2024-01-12 23:29:04.595000', 14),
+       (11, 'd8b6eda1-2ac7-4190-8523-389b3cccff10', 'Yury', 'Petrov', 'Male', 'HB', 5, '2024-01-12 23:29:04.595000',
+        '2024-01-12 23:29:04.595000', 16),
+       (12, 'f4036a98-9fe6-4b85-bf89-3342807872c2', 'John', 'Doe', 'Male', 'AB', 123456, '2024-01-14 20:10:17.988000',
+        '2024-01-14 20:10:17.988000', 22);
