+++ src/main/resources/db/changelog/person_house.sql
@@ -1,0 +1,4 @@
+INSERT INTO public.person_house (person_id, house_id)
+VALUES (1, 1),
+       (3, 3),
+       (12, 21);
